+++ src/models/__init__.py
@@ -1,0 +1,3 @@
+from .auth import ApiKeyCreateRequest
+
+__all__ = ["ApiKeyCreateRequest"]
