+++ src/routes/__init__.py
@@ -1,0 +1,19 @@
+from .auth import router as auth_router
+from .datetime import router as datetime_router
+from .geo import router as geo_router
+from .math import router as math_router
+from .random import router as random_router
+from .root import router as root_router
+from .text import router as text_router
+from .uuid_hashing import router as uuid_hashing_router
+
+__all__ = [
+  "auth_router",
+  "datetime_router",
+  "geo_router",
+  "math_router",
+  "random_router",
+  "root_router",
+  "text_router",
+  "uuid_hashing_router",
+]
